pgdump: Add option for dating filename

+++ scripts/pgdump.sh
@@ -185,6 +185,7 @@ for ((server_idx = 0; server_idx < server_count; server_idx++)); do
     continue
   fi
 
+  filenamedate=$(jq -r ".jobs[${job_idx}].servers[${server_idx}].filenamedate" "${CONFIGFILE}" | sed 's/^null$//g')
   compress=$(jq -r ".jobs[${job_idx}].servers[${server_idx}].compress" "${CONFIGFILE}" | sed 's/^null$//g')
 
   databases_included=$(json_array_to_strlist ".jobs[${job_idx}].servers[${server_idx}].databases_included")
@@ -354,7 +355,11 @@ for ((server_idx = 0; server_idx < server_count; server_idx++)); do
     done
 
     BACKUPFILE_TEMP="${backuppath}/${database}-$(date '+%Y%m%d%H%M%S').tar"
-    BACKUPFILE_FINAL="${backuppath}/${database}.tar"
+    if [ "${filenamedate}" = "true" ]; then
+      BACKUPFILE_FINAL="${BACKUPFILE_TEMP}"
+    else
+      BACKUPFILE_FINAL="${backuppath}/${database}.tar"
+    fi
 
     print "Running pg_dump of ${database} for ${PGHOST} to backupfile ${BACKUPFILE_FINAL}..."
 
@@ -398,12 +403,14 @@ for ((server_idx = 0; server_idx < server_count; server_idx++)); do
       continue
     fi
 
-    mv -v "${BACKUPFILE_TEMP}" "${BACKUPFILE_FINAL}"
-    if [ $? -ne 0 ]; then
-      error "Failed to rename backupfile ${BACKUPFILE_TEMP} to ${BACKUPFILE_FINAL}."
-      rm -f "${BACKUPFILE_TEMP}"
-      result=1
-      continue
+    if ! [ "${BACKUPFILE_TEMP}" = "${BACKUPFILE_FINAL}" ]; then
+      mv -v "${BACKUPFILE_TEMP}" "${BACKUPFILE_FINAL}"
+      if [ $? -ne 0 ]; then
+        error "Failed to rename backupfile ${BACKUPFILE_TEMP} to ${BACKUPFILE_FINAL}."
+        rm -f "${BACKUPFILE_TEMP}"
+        result=1
+        continue
+      fi
     fi
 
     print "Backup of ${database} on ${PGHOST} to backupfile ${BACKUPFILE_FINAL} is successful."
